// src/modules/general/enums/registerLabel.ts
export enum RegisterLabel{
    title = 'Registro',
    name = 'Nombre',
    lastName = 'Apellido',
    email = 'Correo',
    date = 'Fecha de Nacimiento',
    password = 'Contraseña',
    repeatPassword = 'Repetir Contraseña',
    genre = 'Sexo',
    cardTitle = 'Ojo',
    cardInformation = 'Toda la información es verificada por la división de sistemas, cualquier intento de fraude...',
    cancel = 'Cancelar',
    register = 'Registrar'
}